refactor: balanced amounts
